update: add check by casl

+++ backend/src/application/reading-room-interactions/use-cases/remove-reaction/remove-reaction.use-case.ts
@@ -2,6 +2,9 @@ import { Injectable } from '@nestjs/common';
 import { IReactionRepository } from '@/domain/reading-room-interactions/repositories/reaction.repository.interface';
 import { NotFoundDomainException } from '@/shared/domain/common-exceptions';
 import { RemoveReactionCommand } from './remove-reaction.command';
+import { Action, Subject } from '@socialbook/shared';
+import { subject } from '@casl/ability';
+import { ForbiddenDomainException } from '@/shared/domain/common-exceptions';
 
 @Injectable()
 export class RemoveReactionUseCase {
@@ -17,6 +20,12 @@ export class RemoveReactionUseCase {
 
     if (!existing) {
       throw new NotFoundDomainException('Không tìm thấy cảm xúc');
+    }
+
+    if (!command.ability.can(Action.Delete, subject(Subject.RoomReaction, existing))) {
+      throw new ForbiddenDomainException(
+        'You can only remove your own reactions',
+      );
     }
 
     await this.reactionRepository.delete(existing.id);
